+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - support week in date and interval, e.g. `2024-W41-3T06:40+02/W42-7`
 - fix jump leap year every 400 years, heard about it but lacked personal experience
+- expose function to get last week of year
+- expose function to get date for Monday week one
 
 ## [2.0.2] - 2024-09-08